# static/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.0
  title: ElasticSearch Authentication and Authorization Proxy REST API
  description: |
    This is the ElasticSearch AA Proxy REST API documentation.
host: localhost:9000
basePath: /elaap/api/v1
schemes:
  - http
  - https
...
paths:
...
  '/permissions':
    post:
      tags:
        - Permissions
      description: |
        Create a new permission, there can only be one permission object
        per user id / role id combination.
      parameters:
        - name: permission
          in: body
          description: The role JSON representation you want to create
          schema:
            $ref: '#/definitions/Permission:POST'
      responses:
        '201':
          description: Created user object
          schema:
            $ref: '#/definitions/Permission:GET'
        '400':
          description: One or More paramters are invalid
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Role with this ID already exists
          schema:
            $ref: '#/definitions/Error'
...
  '/roles':
    post:
      tags:
        - Roles
      description: |
        Create a new role, roles must be unique.
      parameters:
        - name: role
          in: body
          description: The role JSON representation you want to create
          schema:
            $ref: '#/definitions/Role:POST'
      responses:
        '201':
          description: Created user object
          schema:
            $ref: '#/definitions/Role:GET'
        '400':
          description: One or More paramters are invalid
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Role with this ID already exists
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  'Permission:GET':
    type: object
    properties:
      _id:
        type: string
      description:
        type: string
      permissions:
        type: array
        items: string
      roles:
        type: array
        items: string
      scope:
        type: string
...
  'Permission:POST':
    type: object
    properties:
      _id:
        type: string
      description:
        type: string
      permissions:
        type: array
        items: string
      roles:
        type: array
        items: string
      scope:
        type: string
...
  'Error':
    type: object
    properties:
      err:
        type: string
  'Role:GET':
    type: object
    properties:
      _id:
        type: string
      description:
        type: string
      users:
        type: array
        items: string
...
  'Role:POST':
    type: object
    properties:
      _id:
        type: string
      description:
        type: string
      users:
        type: array
        items: string
...